+++ NEWS.md
@@ -1,4 +1,6 @@
 # occumb (development version)
+* Change occumbData class specification and occumbData() function.
+* Add occumb() function.
 
 # occumb v0.1.0
 * Initial development of occumb package.